refactor normalization utils
